Research endpoint changes

+++ src/main/kotlin/pl/umk/mat/zesp01/pz2022/researcher/controller/ResearchController.kt
@@ -90,6 +90,27 @@ class ResearchController(
 		}
 	}
 
+//	@GetMapping("/research/{code}/enrollCheck")
+//	fun checkCurrentUserEnrollment(
+//		@PathVariable code: String,
+//		@RequestHeader httpHeaders: HttpHeaders
+//	): ResponseEntity<String> {
+//		val jwt = httpHeaders["Authorization"]
+//			?: return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build()
+//
+//		return try {
+//			val username = refreshTokenService.verifyAccessToken(jwt[0]) ?: throw Exception()
+//			if (username.isEmpty()) throw Exception()
+//
+//			if (researchService.checkIfUserIsAlreadyOnParticipantsList(code, username))
+//				ResponseEntity.status(HttpStatus.OK).build()
+//			else ResponseEntity.status(HttpStatus.NO_CONTENT).build()
+//
+//		} catch (e: Exception) {
+//			ResponseEntity.status(HttpStatus.FORBIDDEN).build()
+//		}
+//	}
+
 	@GetMapping("/research/all")
 	fun getAllResearches(): ResponseEntity<List<Research>> =
 		ResponseEntity.status(HttpStatus.OK).body(researchService.getAllResearches())
@@ -152,11 +173,13 @@ class ResearchController(
 		val jwt = httpHeaders["Authorization"]
 			?: return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build()
 
-		// usuwac badanie JEDYNIE gdy jest to badanie zalogowanego uzytkownika
-
 		return try {
 			val username = refreshTokenService.verifyAccessToken(jwt[0]) ?: throw Exception()
 			if (username.isEmpty()) throw Exception()
+
+			// You can delete a research only if it's logged user's research
+			if(!researchService.isGivenUserAResearchCreator(code, username))
+				return ResponseEntity.status(HttpStatus.FORBIDDEN).build()
 
 			researchService.deleteResearchByResearchCode(code)
 			ResponseEntity.status(HttpStatus.NO_CONTENT).build()
